+++ src/main/kotlin/com/github/sleepypanda/feesh/FeeshMod.kt
@@ -132,6 +132,7 @@ class FeeshMod : ClientModInitializer {
         WaterHotspotsTracker.init()
         CrimsonIsleTracker.init()
         GalateaWaterTracker.init()
+        LotusAtollTracker.init()
         ArchfiendDiceProfitTracker.init()
         FishingFestivalTracker.init()
         FishingProfitTracker.init()