+++ src/main/kotlin/com/github/sleepypanda/feesh/utils/data/FeeshData.kt
@@ -7,6 +7,7 @@ import com.github.sleepypanda.feesh.features.overlays.BayouTracker
 import com.github.sleepypanda.feesh.features.overlays.WaterHotspotsTracker
 import com.github.sleepypanda.feesh.features.overlays.CrimsonIsleTracker
 import com.github.sleepypanda.feesh.features.overlays.GalateaWaterTracker
+import com.github.sleepypanda.feesh.features.overlays.LotusAtollTracker
 import com.github.sleepypanda.feesh.features.overlays.ArchfiendDiceProfitTracker
 import com.github.sleepypanda.feesh.features.overlays.FishingProfitTracker
 import com.github.sleepypanda.feesh.features.overlays.MagmaCoreFishingTracker
@@ -23,6 +24,7 @@ data class FeeshData(
     val waterHotspotsTracker: WaterHotspotsTracker.WaterHotspotsTrackerData = WaterHotspotsTracker.WaterHotspotsTrackerData(),
     val crimsonIsle: CrimsonIsleTracker.CrimsonIsleTrackerData = CrimsonIsleTracker.CrimsonIsleTrackerData(),
     val galateaWater: GalateaWaterTracker.GalateaWaterTrackerData = GalateaWaterTracker.GalateaWaterTrackerData(),
+    val lotusAtollTracker: LotusAtollTracker.LotusAtollTrackerData = LotusAtollTracker.LotusAtollTrackerData(),
     val archfiendDiceProfit: ArchfiendDiceProfitTracker.ArchfiendDiceProfitData = ArchfiendDiceProfitTracker.ArchfiendDiceProfitData(),
     val fishingProfit: FishingProfitTracker.FishingProfitData = FishingProfitTracker.FishingProfitData(),
     val magmaCoreFishing: MagmaCoreFishingTracker.MagmaCoreFishingData = MagmaCoreFishingTracker.MagmaCoreFishingData(),